Add user-persistence for upvotes/downvotes and start user_answers_index
