+++ frontend/src/components/answer/user_answer_index_container.js
@@ -1,0 +1,20 @@
+import { connect } from 'react-redux';
+import { fetchAnswers, updateAnswer, deleteAnswer } from '../../actions/answer_actions';
+import AnswerIndex from './user_answer_index';
+// import AnswerIndexStylesheet from '../../../stylesheets/user_answer_index.scss';
+
+const mapStateToProps = state => ({
+    answers: Object.values(state.entities.answers),
+    currentUser: state.session.user
+});
+
+const mapDispatchToProps = dispatch => ({
+    fetchAnswers: optionId => dispatch(fetchAnswers(optionId)),
+    updateAnswer: answer => dispatch(updateAnswer(answer)),
+    deleteAnswer: id => dispatch(deleteAnswer(id))
+});
+
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps)
+(AnswerIndex);
