+++ frontend/src/components/question/answers/question_answer_index.jsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import { withRouter } from 'react-router-dom';
+import AnswerIndexItem from './question_answer_index_item';
+
+class AnswerIndex extends React.Component {
+    componentDidMount() {
+        this.props.fetchAnswers({questionId: this.props.match.params.question_id});
+    }
+
+    render() {
+        const answers = this.props.answers.map((answer, idx) => 
+            <AnswerIndexItem
+                key={`answer-${idx}`}
+                answer={answer}
+                {...this.props} />
+        );
+
+        return (<div className='question-answer-index-container'>
+            <div className='answer-count'>
+                {answers.length} Answers
+            </div>
+            {answers}
+        </div>)
+    }
+}
+
+export default withRouter(AnswerIndex);
