+++ frontend/src/stylesheets/question_answer_index.scss
@@ -1,0 +1,154 @@
+.question-answer-index-container {
+    width: 600px;
+    min-width: 600px;
+
+    .answer-count {
+        font-size: 18px;
+        font-weight: 500;
+    }
+
+    .content-divider {
+        width: 100%;
+        height: 1px;
+        background: #e2e2e2;
+        margin: 8px 0px 12px 0px;
+    }
+
+    .answer-header {
+        margin-bottom: 8px;
+
+        .answer-author {
+            font-size: 15px;
+            margin-bottom: 2px;
+            cursor: pointer;
+
+            &:hover {
+                text-decoration: underline;
+            }
+        }
+    
+        .answer-date {
+            color: #949494;
+            cursor: pointer;
+    
+            &:hover {
+                text-decoration: underline;
+                text-decoration-color: #949494;
+            }
+        }
+    }
+
+    .answer-body {
+        font-size: 16px;
+    }
+
+    .answer-footer {
+        display: flex;
+        justify-content: space-between;
+        font-size: 14px;
+
+        .svg-icon {
+            line-height: 0px;
+            vertical-align: middle;
+
+            .svg-base {
+                stroke: #666;
+                stroke-width: 1.5;
+                stroke-linejoin: round;
+                fill: none;
+                fill-rule: evenodd;
+                transition: all 100ms ease-in-out 0s;
+            }
+        }
+
+        .answer-upvote {
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+            color: #666;
+            font-size: 14px;
+            padding: 7px 12px;
+            border-radius: 40px;
+            cursor: pointer;
+
+            .svg-base {
+                stroke: #329bff;
+            }
+
+            .upvote-txt {
+                font-weight: 500;
+                margin: 0px 5px;
+            }
+
+            .upvote-bullet {
+                margin-right: 5px;
+                opacity: 0.5;
+            }
+
+            &:hover {
+                background: rgba(0, 0, 0, 0.02);
+            }
+
+            &.active {
+                color: #329bff;
+
+                .svg-base {
+                    fill: #329bff;
+                }
+            }
+        }
+
+        .answer-downvote {
+            display: flex;
+            align-items: center;
+            padding: 7px;
+            border-radius: 40px;
+            cursor: pointer;
+
+            &:hover {
+                background: rgba(0, 0, 0, 0.02);
+            }
+
+            &.active {
+                .svg-base {
+                    stroke: #329bff;
+                    fill: #329bff;
+                }
+            }
+
+            &.tooltip {
+                position: relative;
+
+                &::before {
+                    background-color: #444;
+                    content: ' ';
+                    position: absolute;
+                    top: 0;
+                    left: 50%;
+                    transform: translate(-50%, calc(-100% + 5px)) rotate(45deg);
+                    width: 7px;
+                    height: 7px;
+                    z-index: 999;
+                }
+    
+                &::after {
+                    position: absolute;
+                    top: 0px;
+                    left: 50%;
+                    transform: translate(-50%, calc(-100% + 2px));
+                    content: 'Downvote';
+                    color: white;
+                    font-size: 12px;
+                    line-height: 12px;
+                    text-align: center;
+                    padding: 7px;
+                    background: #444;
+                    border-radius: 3px;
+                    z-index: 999;
+                }
+            }
+
+        }
+    }
+
+}
