+++ frontend/src/stylesheets/session_form.scss
@@ -17,7 +17,7 @@
     height: 100vh;
     width: 100%;
     box-sizing: content-box;
-    background-image: url('../images/BeFunky-collage.png');
+    background-image: url('../images/BeFunky-collage-expanded.png');
     background-size: cover;
     background-repeat: no-repeat;
     background-position: center center;
