+++ routes/api/answers.js
@@ -5,14 +5,14 @@ const passport = require('passport');
 
 const Answer = require('../../models/Answer');
 
-router.get("/test", (req, res) => res.json({ answer: "Answers Route" }));
+router.get('/test', (req, res) => res.json({ answer: 'Answers Route' }));
 
 // find by answer_id
-router.get("/:answer_id", (req, res) => {
+router.get('/:answer_id', (req, res) => {
     Answer.findById(req.params.answer_id)
         .then(answer => res.json(answer))
         .catch(err =>
-            res.status(404).json({ answer: "No answer found with that ID" })
+            res.status(404).json({ answer: 'No answer found with that ID' })
         );
 });
 
@@ -21,10 +21,10 @@ router.patch('/:answer_id',
     passport.authenticate('jwt', { session: false }), (req, res) => {
         Answer.findOneAndUpdate({ _id: req.params.answer_id },
         {
-            $set:
-            {
+            $set: {
                 upvote: req.body.upvote,
                 downvote: req.body.downvote,
+                voters: req.body.voters,
                 description: req.body.description
             }
         },
